Add test cases for part 2
The assignment specifies that the data should be read in from a file, so
I used a structured format to make it easy to parse. The test simply
evaluates the expressions under all the provided assignments.

TODO: Add 6 more test cases as specified. These will automatically be
run thanks to the way the test is set up.